+++ daemon/remote/XmlRpc.h
@@ -107,7 +107,7 @@ protected:
 
 	void BuildErrorResponse(int errCode, const char* errText, ...);
 	void BuildBoolResponse(bool ok);
-	void BuildIntResponse(int value);
+	void BuildIntResponse(int64 value);
 	void AppendResponse(const char* part);
 	void AppendFmtResponse(const char* format, ...);
 	void AppendCondResponse(const char* part, bool cond);
@@ -115,7 +115,7 @@ protected:
 	bool NextParamAsInt(int64* value);
 	bool NextParamAsBool(bool* value);
 	bool NextParamAsStr(char** valueBuf);
-	char* XmlNextValue(char* xml, const char* tag, int* valueLength);
+	char* XmlNextValue(char* xml, const char* tag, int64* valueLength);
 	const char* BoolToStr(bool value);
 	CString EncodeStr(const char* str);
 	void DecodeStr(char* str);
